Commit Feb 23 2023

+++ create_pipeline_performancetable.sh
@@ -359,7 +359,7 @@ sessionID_1="${array[1]}"
 echo final_output_directory::${final_output_directory}
 niftifile_csvfilename=${working_dir}/${sessionID_1}'this_session_final_ct.csv'
 get_nifti_scan_uri ${sessionID_1}  ${working_dir} ${niftifile_csvfilename}
-if [ -f ${niftifile_csvfilename} ]; then
+if [ -f ${niftifile_csvfilename} ] && [ $counter -lt 1 ]; then
     echo "$niftifile_csvfilename exists."
     cp ${niftifile_csvfilename} ${final_output_directory}
     #############
@@ -393,7 +393,7 @@ if [ -f ${niftifile_csvfilename} ]; then
 
     ###################
 
-#    counter=$((counter+1))
+    counter=$((counter+1))
 fi
 #if [[ $counter -gt 5 ]] ; then
 #  break
